fix(ventas): cantidades decimales, descuento cambio de tipo, LPN del ticket, rollback venta sin líneas — v0.85.2
- Migration 054: venta_items.cantidad INT → DECIMAL(14,4) para soportar kg/l/etc
- Descuento cambio de tipo ($ ↔ %): clampea valor al nuevo límite en vez de dejar nro inválido
- LPN eliminado del ticket del cliente (dato interno de almacén)
- Cantidad decimal: defaultValue con coma; UOM enteras bloquean punto/coma en onKeyDown
- Stock guard: solo valida si lineas_disponibles.length > 0 (fix post-restore carrito)
- registrarVenta: valida cantidad antes de insertar + rollback DELETE header si items fallan
- esDecimal + parseCantidad extraídas a ventasValidation.ts — 24 tests nuevos (178/178 passing)

+++ supabase/schema_full.sql
@@ -483,7 +483,7 @@ CREATE TABLE venta_items (
   venta_id               UUID NOT NULL REFERENCES ventas(id) ON DELETE CASCADE,
   producto_id            UUID NOT NULL REFERENCES productos(id),
   linea_id               UUID REFERENCES inventario_lineas(id),
-  cantidad               INT NOT NULL DEFAULT 1,
+  cantidad               DECIMAL(14,4) NOT NULL DEFAULT 1,
   precio_unitario        DECIMAL(12,2) NOT NULL,
   descuento              DECIMAL(5,2) NOT NULL DEFAULT 0,
   subtotal               DECIMAL(12,2) NOT NULL,
